pyproject.toml
[tool.poetry]
name = "DeepSurveySim"
version = "0.1.0"
description = "An astrological survey simulation designed for testing MDP-style algorithms"
authors = ["M. Voetberg <maggiev@fnal.gov"]
license = "BSD 3-Clause"
readme = "README.md"
homepage = "https://github.com/deepskies/DeepSurveySim"
repository = "https://github.com/deepskies/DeepSurveySim"
documentation =  "https://deepsurveysim.readthedocs.io/en/latest/"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
numpy = "^1.24.2"
pandas = "^1.5.3"
astroplan = "^0.8"
PyYAML = "^6.0"
numexpr = "^2.8.4"
skybright = {git = "https://github.com/ehneilsen/skybright.git", optional = true}
configparser = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"
pre-commit = "^3.3.2"
sphinx = "^7.0.1"
sphinx-autodoc-typehints = "^1.23.3"

[tool.poetry.extras]
skybright = ["skybright"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
